+++ DataAccessLayer/Models/Product.cs
@@ -16,7 +16,7 @@ namespace DataAccessObject.Models
         public int Id { get; set; }
         public string ProductName { get; set; }
         public string? Description { get; set; }
-        public double ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int? Quantity { get; set; }
         public string? MadeIn { get; set; }
         public string? ShipFrom { get; set; }
